+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Low blood sugar no longer affects sleep if sleep restrictions are disabled
 - Experiments can now estimate the amount of carbohydrate in foods and drinks; their ability to do so improves with INT
 - Injecting sugary liquids will directly increase blood sugar
+- Electronic watches periodically report your blood sugar
 
 ## 1.1.0
 - Game no longer crashes when Casualties: Together isn't installed